For the Hoe's
I gave the druids a stone hoe. which notoriously breaks after three uses. They dont really need it, but it seems to make some sense.

Added some more thorns and details to the paths, and moved some things to make the line of sight and flow of the base feel a bit less confusing on the eyes